requirements.txt
Flask==2.3.3
openai==0.28.1
python-dotenv==1.0.0
pandas==2.0.3
gunicorn==21.2.0